修改保存偏移量的地址
